add some scripts that try to make PowerSim more efficient.
powerSim() is a terribly slow process that will only provide results for a single test at a time. I've written a version that will test for main effects of two variables and an interaction at the same time. It's limited in the range of models and tests you can run, but at least it saves you running the simulations repeatedly.